+++ src/main/java/com/synergy/backend/domain/member/repository/AttendeeRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.synergy.backend.domain.member.repository;
 
-import com.synergy.backend.domain.member.api.dto.AttendeeFilterRequest;
+import com.synergy.backend.domain.member.api.dto.request.AttendeeFilterRequest;
 import com.synergy.backend.domain.member.api.dto.resposne.AttendeeSimpleResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
