+++ src/main/java/com/synergy/backend/domain/member/service/RecruiterService.java
@@ -1,7 +1,7 @@
 package com.synergy.backend.domain.member.service;
 
-import com.synergy.backend.domain.member.api.dto.AttendeeFilterRequest;
-import com.synergy.backend.domain.member.api.dto.AttendeeListResponse;
+import com.synergy.backend.domain.member.api.dto.request.AttendeeFilterRequest;
+import com.synergy.backend.domain.member.api.dto.resposne.AttendeeListResponse;
 import com.synergy.backend.domain.member.api.dto.resposne.RecruiterMyInfoResponseDto;
 import org.springframework.data.domain.Pageable;
